updated project structure 0.0.1
